+++ integration_test/scenarios/task_scenario.dart
@@ -13,9 +13,9 @@ import 'package:mem/logger/log.dart';
 import 'package:mem/logger/log_service.dart';
 import 'package:mem/notifications/notification_client.dart';
 import 'package:mem/notifications/notification/type.dart';
-import 'package:mem/settings/client.dart';
-import 'package:mem/settings/keys.dart';
-import 'package:mem/settings/preference.dart';
+import 'package:mem/settings/preference/client.dart';
+import 'package:mem/settings/preference/keys.dart';
+import 'package:mem/settings/preference/preference.dart';
 
 import 'helpers.dart';
