+++ lib/settings/page.dart
@@ -1,14 +1,16 @@
 import 'package:flutter/material.dart';
 import 'package:flutter_riverpod/flutter_riverpod.dart';
+import 'package:settings_ui/settings_ui.dart';
+
 import 'package:mem/l10n/l10n.dart';
 import 'package:mem/logger/log_service.dart';
 import 'package:mem/notifications/notification_client.dart';
-import 'package:mem/settings/actions.dart';
-import 'package:mem/settings/backup_section.dart';
-import 'package:mem/settings/keys.dart';
-import 'package:mem/settings/states.dart';
 import 'package:mem/values/constants.dart';
-import 'package:settings_ui/settings_ui.dart';
+
+import 'actions.dart';
+import 'backup_section.dart';
+import 'preference/keys.dart';
+import 'states.dart';
 
 class SettingsPage extends ConsumerWidget {
   const SettingsPage({super.key});
