+++ lib/settings/preference/client.dart
@@ -1,7 +1,7 @@
 import 'package:mem/framework/repository/key_with_value_repository.dart';
 import 'package:mem/logger/log_service.dart';
-import 'package:mem/settings/preference.dart';
-import 'package:mem/settings/preference_key.dart';
+import 'package:mem/settings/preference/preference.dart';
+import 'package:mem/settings/preference/preference_key.dart';
 import 'package:shared_preferences/shared_preferences.dart';
 
 class PreferenceClientRepository
