+++ lib/settings/preference/keys.dart
@@ -1,3 +1,3 @@
-import 'package:mem/settings/preference_key.dart';
+import 'package:mem/settings/preference/preference_key.dart';
 
 final startOfDayKey = TimeOfDayPreferenceKey("start_of_day");
