+++ lib/settings/preference/preference.dart
@@ -1,6 +1,6 @@
 import 'package:mem/framework/repository/entity.dart';
 import 'package:mem/framework/repository/key_with_value.dart';
-import 'package:mem/settings/preference_key.dart';
+import 'package:mem/settings/preference/preference_key.dart';
 
 class PreferenceEntity<VALUE>
     with Entity, KeyWithValue<PreferenceKey<VALUE>, VALUE?> {
