+++ lib/settings/states.dart
@@ -1,10 +1,12 @@
 import 'package:flutter/material.dart';
 import 'package:flutter_riverpod/flutter_riverpod.dart';
+
 import 'package:mem/framework/view/value_state_notifier.dart';
 import 'package:mem/logger/log_service.dart';
-import 'package:mem/settings/actions.dart';
-import 'package:mem/settings/keys.dart';
 import 'package:mem/values/constants.dart';
+
+import 'actions.dart';
+import 'preference/keys.dart';
 
 final startOfDayProvider =
     StateNotifierProvider<ValueStateNotifier<TimeOfDay>, TimeOfDay>(
